Hotfixes 4 (#43)
* U game_data.h: removed duplicate isMult

* U gameobject: fixed pointy stuff

* U drawing.h: re-added const, removed virtual

doesn't need to be virtual, every subclass has the same get_x/y methods

* U button: got rid of some compiler warnings

* U GObjFactory: re-added atLeastOne

* U GObjFactory.h: same as GObjFactory.cpp

also fixed the cast lmao

+++ button.cpp
@@ -13,7 +13,7 @@ void Button::update(float not_used_lmao)
 		graphics::getMouseState(ms);
 		if (ms.button_left_pressed)
 		{
-			clicked = (((abs(mouse_x(ms.cur_pos_x) - x) <= width)) && (abs(mouse_y(ms.cur_pos_y) - y) <= height));
+			clicked = (((abs(mouse_x((float)ms.cur_pos_x) - x) <= width)) && (abs(mouse_y((float)ms.cur_pos_y) - y) <= height));
 		}
 	}
 }
